Improve restartability of the mapping phase.

+++ mapping/revert-bam.sh
@@ -85,6 +85,7 @@ fi
 if [ -e $reference ];
 then
   echo "Mapping, marking duplicates and chromosome order sorting $outfile into ${bamfile%%.bam}.srt.bam"
+
   # Identify Complete Genomics style read names for library complexity estimation
   try_regex='CL10.......L.C([0-9]+)R([0-9]+)_([0-9]+)'
   if samtools view $outfile | head -1 | cut -f1 | grep -q -E "$try_regex";
@@ -92,16 +93,24 @@ then
     regex="READ_NAME_REGEX=\"$try_regex\""
   fi
 
-  samtools view -H $outfile > ${outfile}.hdr
-  samtools fastq -t $outfile \
-    | bwa mem -p -t $cores -M -C -H ${outfile}.hdr $reference - \
-    | samtools view -b -o ${bamfile%%.bam}.mem.bam
-  rm ${outfile}.hdr
+  # Uses .hdr file also as a flag of whether mapping finished, in case we restart
+  if [ ! -e ${bamfile%%.bam}.mem.bam ] && [ ! -e ${outfile}.hdr ];
+  then
+    samtools view -H $outfile > ${outfile}.hdr
+    samtools fastq -t $outfile \
+      | bwa mem -p -t $cores -M -C -H ${outfile}.hdr $reference - \
+      | samtools view -b -o ${bamfile%%.bam}.mem.bam
+    rm ${outfile}.hdr
+  fi
+
   # Unfortunately, MarkDuplicates seeks back to beginning of the input BAM so mapping can't just be piped in
-  java -jar picard.jar MarkDuplicates INPUT=${bamfile%%.bam}.mem.bam OUTPUT=/dev/stdout METRICS_FILE=${bamfile}.dup \
-    ASSUME_SORT_ORDER=queryname TAGGIN_POLICY=All COMPRESSION_LEVEL=0 TMP_DIR=$tmp \
-    OPTICAL_DUPLICATE_PIXEL_DISTANCE=2500 $regex \
-    | samtools sort -T $tmp/$bamfile -@$cores -m${coremem}G -l9 -o ${bamfile%%.bam}.srt.bam
+  (
+    java -jar picard.jar MarkDuplicates INPUT=${bamfile%%.bam}.mem.bam OUTPUT=/dev/stdout METRICS_FILE=${bamfile}.dup \
+      ASSUME_SORT_ORDER=queryname TAGGING_POLICY=All COMPRESSION_LEVEL=0 TMP_DIR=$tmp \
+      OPTICAL_DUPLICATE_PIXEL_DISTANCE=2500 $regex \
+        | samtools sort -T $tmp/$bamfile -@$cores -m${coremem}G -l9 -o ${bamfile%%.bam}.srt.bam
+  )
+  # Check result of previous subshell, set -e and set -o pipefail not an option without bash
   if [ $? -eq 0 ];
   then
     rm ${bamfile%%.bam}.mem.bam
